requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
yt-dlp>=2023.11.16
tenacity>=8.2.3